requirements.txt
openai
faiss-cpu
streamlit
python-dotenv
pytest
huggingface-hub
sentence-transformers
ollama